ovn: move master subnet allocator to its own file
(cherry picked from commit 3161d148c2ccfe445057c1084640eb4f0eff82ee)
(cherry picked from commit 302b3080fcca4ca149ed64e1ed49c432e2d596bf)
(cherry picked from commit 2d28afb74a330515f121e776f94de60b1000c1fe)
